test: hand out a fixture only once it is genuinely walkable
A pumping window is necessary but not sufficient. ElementFromHandle sends
WM_GETOBJECT and that SendMessage carries its own timeout, so on a loaded
runner starting several fixture hosts at once the first call returns E_FAIL
against a window that is healthy a moment later. The fixture's contract is a
walkable window, so it does not hand one out until it is one - which keeps the
retry in the harness rather than smuggling it into the product's resolver.

Also splits the live property tests out of properties_tests.rs, which was at
377 of the 400-line cap.
