+++ crates/windows/src/tree/properties_live_tests.rs
@@ -1,0 +1,169 @@
+use super::*;
+use crate::tree::fixture::{CONTENT_MARKER, HostedFixture, SECURE_MARKER};
+use crate::tree::properties::read_cached;
+use uiautomation::UIElement;
+
+fn walk_children(handle: isize) -> Vec<crate::tree::element::UIAElement> {
+    let root = crate::tree::automation::root_from_hwnd(
+        handle,
+        agent_desktop_core::Deadline::standard().expect("a standard deadline"),
+    )
+    .expect("the fixture window resolves");
+    let client = crate::tree::automation::automation_client().expect("a UIA client");
+    let walker = client
+        .get_raw_view_walker()
+        .expect("the raw view walker is available");
+    let mut children: Vec<UIElement> = Vec::new();
+    if let Ok(mut current) = walker.get_first_child(&root.0) {
+        loop {
+            let next = walker.get_next_sibling(&current);
+            children.push(current);
+            match next {
+                Ok(sibling) => current = sibling,
+                Err(_) => break,
+            }
+        }
+    }
+    children
+        .into_iter()
+        .map(crate::tree::element::UIAElement::from)
+        .collect()
+}
+
+fn bootstrap() {
+    crate::tree::fixture::ensure_test_apartment();
+}
+
+/// The gate asserted against a real out-of-process provider, not only
+/// against the projection: text typed into the fixture's `ES_PASSWORD`
+/// control must appear in no read outcome for `Value`, `Name` or
+/// `HelpText`.
+#[test]
+fn secure_content_never_reaches_a_read_outcome_from_a_live_provider() {
+    bootstrap();
+    let fixture = HostedFixture::spawn().expect("the fixture host starts");
+    let children = walk_children(fixture.handle());
+
+    let secure = children
+        .iter()
+        .map(|child| read_live(child).0)
+        .find(ElementProperties::is_secure)
+        .expect("the fixture exposes a control reporting IsPassword");
+
+    for property in TreeProperty::VALUE_BEARING {
+        let rendered = format!("{:?}", secure.get(property));
+        assert!(
+            !rendered.contains(SECURE_MARKER),
+            "{} leaked secure content",
+            property.as_str()
+        );
+        assert_eq!(secure.get(property), PropertyOutcome::Absent);
+    }
+}
+
+#[test]
+fn a_live_read_distinguishes_present_content_from_an_unimplemented_property() {
+    bootstrap();
+    let fixture = HostedFixture::spawn().expect("the fixture host starts");
+    let children = walk_children(fixture.handle());
+    let all: Vec<ElementProperties> = children.iter().map(|child| read_live(child).0).collect();
+
+    assert!(
+        all.iter().any(|properties| matches!(
+            properties.get(TreeProperty::Value),
+            PropertyOutcome::Known(PropertyValue::Text(ref value)) if value.contains(CONTENT_MARKER)
+        )),
+        "a plain control's content must read back as Known"
+    );
+    assert!(
+        all.iter().any(|properties| matches!(
+            properties.get(TreeProperty::AutomationId),
+            PropertyOutcome::Known(_) | PropertyOutcome::Absent
+        )),
+        "an unimplemented property must classify Absent, never Unknown"
+    );
+    assert!(
+        all.iter().all(|properties| !matches!(
+            properties.get(TreeProperty::ClassName),
+            PropertyOutcome::Unknown
+        )),
+        "a readable property must never classify Unknown on a healthy provider"
+    );
+}
+
+/// A14-9: once the host process exits, the client-side HWND proxy answers
+/// property reads locally with an empty string rather than failing, so
+/// process death is no more detectable through a property read than
+/// through the sibling terminator (A14-4).
+///
+/// The guarantee 2.2 does make is asserted instead, and it is the one
+/// that matters downstream: a provider that went away is never reported
+/// as a provider that does not implement the property. `Absent` feeds
+/// completeness gating as a legitimate answer; fabricating it here would
+/// let a dead target satisfy `EvidenceRequirements` it never answered.
+#[test]
+fn a_read_after_the_provider_exits_never_fabricates_absent() {
+    bootstrap();
+    let mut fixture = HostedFixture::spawn().expect("the fixture host starts");
+    let children = walk_children(fixture.handle());
+    assert!(!children.is_empty(), "the fixture exposes child controls");
+
+    fixture.terminate();
+    std::thread::sleep(std::time::Duration::from_millis(750));
+
+    let (properties, _) = read_live(&children[0]);
+
+    for property in [
+        TreeProperty::ClassName,
+        TreeProperty::Name,
+        TreeProperty::Value,
+    ] {
+        assert_ne!(
+            properties.get(property),
+            PropertyOutcome::Absent,
+            "a provider that went away must not be reported as not implementing {}",
+            property.as_str()
+        );
+    }
+}
+
+/// The failure branch, forced deterministically rather than by killing a
+/// process: a property the cache request never carried cannot be answered,
+/// and must classify `Unknown` with a structured error.
+#[test]
+fn a_read_that_genuinely_fails_classifies_unknown_and_never_absent() {
+    bootstrap();
+    let fixture = HostedFixture::spawn().expect("the fixture host starts");
+    let root = crate::tree::automation::root_from_hwnd(
+        fixture.handle(),
+        agent_desktop_core::Deadline::standard().expect("a standard deadline"),
+    )
+    .expect("the fixture window resolves");
+    let client = crate::tree::automation::automation_client().expect("a UIA client");
+    let request = client
+        .create_cache_request()
+        .expect("an empty cache request builds");
+    request
+        .set_element_mode(uiautomation::types::ElementMode::Full)
+        .expect("the element mode is settable");
+    let sparse = root
+        .0
+        .build_updated_cache(&request)
+        .map(crate::tree::element::UIAElement::from)
+        .expect("an element with an empty cache");
+
+    let (properties, errors) = read_cached(&sparse);
+
+    assert!(
+        !errors.is_empty(),
+        "a read that cannot be answered must surface a structured error"
+    );
+    assert_eq!(
+        properties.get(TreeProperty::ClassName),
+        PropertyOutcome::Unknown
+    );
+    assert_ne!(
+        properties.get(TreeProperty::ClassName),
+        PropertyOutcome::Absent
+    );
+}
